backend/requirements.txt
Flask==3.0.3
Flask_Caching==2.3.0
Flask_Cors==4.0.1
Flask_SocketIO==5.3.6
langchain==0.2.12
langchain_community==0.2.11
langchain_core==0.2.28
langchain_google_genai==1.0.8
langgraph==0.1.19
pymongo==4.7.3
python-dotenv==1.0.1
Requests==2.32.3
typing_extensions==4.12.2
langchain_openai
gevent==24.2.1